% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ModelID.R
\name{get_ModelID}
\alias{get_ModelID}
\title{Get ModelID from a stripped cell line name}
\usage{
get_ModelID(cell_line_name)
}
\arguments{
\item{cell_line_name}{a characeter string of a cell line name matching to an entry in the "StrippedCellLineName" column in the DepMap's Model database.}
}
\value{
a character string correponding to \code{cell_line_name}.
}
\description{
Retrieve the \code{ModelID} that corresponds to a stripped cell line name in the DepMap model database.
}
\examples{
get_ModelID('THP1')

}
\author{
Chao-Jen Wong
}
