% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/per_protein_abun.R
\name{per_protein_abun}
\alias{per_protein_abun}
\title{Simple point plot to show per-protein aboundance}
\usage{
per_protein_abun(se, protein_id, category_by)
}
\arguments{
\item{se}{SummarizedExperience instance (SE) or list of SE containing the transformed protein abundance.}

\item{protein_id}{Protein ID}

\item{group_by}{Group by a factor in the meta file}
}
\description{
Simple point plot to show per-protein aboundance
}
\author{
Chao-Jen Wong
}
