% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Volcano plots for TalusResults}
\usage{
plot_volcano(
  res,
  alpha = 0.05,
  lfc_threshold = 0.5,
  use_adjP = TRUE,
  contrast_levels = NULL,
  label_top_n = 10,
  return_data = FALSE
)
}
\arguments{
\item{res}{a \code{TalusResults} instance containing results returned by \code{talus_limma} or \code{talus_row_t_walch}.}

\item{alpha}{the significance adjust p-vluae cutoff for independent filtering. Default to 0.05.}

\item{lfc_threshold}{a non-negative values of significance log fold change threshold for independent filtering. Default to 0.5.}

\item{use_adjP}{logical to indicate wether to use 'adj.P.val' or 'P.Value'. Default to TRUE.}

\item{contrast_levels}{Default to NULL.}

\item{return_data}{logical to return data only with mutated \code{logp} and \code{sig} columns}
}
\description{
Valcano plots for TalusResults object
}
\author{
Chao-Jen Wong
}
