% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_talus.R
\name{read_talus}
\alias{read_talus}
\title{Import Talus files}
\usage{
read_talus(
  file,
  meta_file,
  which_proteinid = "Protein.Ids",
  which_fraction = "Frx",
  which_sequence = NA,
  which_run = "Run",
  remove_few_measurements = TRUE,
  split_by_fraction = TRUE,
  log_transform = TRUE,
  rowname_repair = TRUE
)
}
\arguments{
\item{file}{A path to a Diann summarized protein file or a connection.}

\item{remove_few_measurements}{If \code{TRUE}, remove measurements that are absent 85\% across all samples}

\item{split_by_fraction}{If \code{TRUE}, split the protein abundance file according to value indicating in the values of the \code{which_fraction} column in the meta file}

\item{log_transform}{If \code{TRUE}, perform log2 transformation to the protein abundance}

\item{meta_data}{A path to a meta (column) data file for samples.}
}
\value{
A \code{SummarizedExperiment} objects with three layers of assay data
}
\description{
Import Talus files that were processed by DIANN and convert to
SummarizedExperiment instance. Can split the data into three instancs
containing fractions in nuc, cyto, and plasm.
}
