% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/talus_limma.R
\name{talus_limma}
\alias{talus_limma}
\title{Wrapper function of limma for differential analysis}
\usage{
talus_limma(se, design = ~0 + Tx)
}
\arguments{
\item{se}{a \code{SummarizedExperiment} instance containing protein log-transform assays}

\item{desgin}{a formula to create a model matrix. Default to ~0 + Tx, where \code{Tx} is a factor from colData of \code{se}}
}
\description{
A wrapper function using lmFit() to perform differential analysis
}
\author{
Chao-Jen Wong
}
