% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{TalusDataSetList}
\alias{TalusDataSetList}
\alias{TalusDataSetList-class}
\title{TalusDataSetList object and constructors}
\usage{
TalusDataSetList(...)
}
\value{
a \code{TalusDataSetList} object.
}
\description{
\code{TalusDataSetList}
}
\author{
Chao-Jen Wong
}
