% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/AllMethods.R
\docType{class}
\name{TalusDataSet-class}
\alias{TalusDataSet-class}
\alias{TalusDataSet}
\alias{show,TalusDataSet-method}
\alias{show,TalusDataSetList-method}
\title{TalusDataSet object and constructor}
\usage{
TalusDataSet(
  assay_data,
  col_data,
  row_data,
  intensity_group = "protein",
  metric = "DIA-NN",
  log_transform = "log2"
)

\S4method{show}{TalusDataSet}(object)

\S4method{show}{TalusDataSetList}(object)
}
\arguments{
\item{assay_data}{a matrix input: a matrix of non-negative integers of MS signal levels}

\item{col_data}{for matrix input: a \code{DataFrame} or \code{data.frame} with at least a single column. Rows of \code{col_data} correspond to columns of \code{assay_data}.}

\item{row_data}{for matrix input: a \code{DataFrame} or \code{data.frame} with at least a single column. Rows of \code{row_data} correspond to the rows of \code{assay_data}.}

\item{intensity_group}{a character string specifying the levels of abundance, either protein or peptide.}

\item{log_transform}{a character string of the log transformation method.}

\item{matric}{a character string specifying the metric that generates the mass-spectrum signal values.}
}
\value{
a \code{TalusDataSet} object.
}
\description{
\code{TalusDataSet} is a subclass of \code{SummerizedExperiment} used
to store the input mass-spectrum intensity signals, which reflects the
protein abundance, and addition annotation of intensity group
(either protein or peptide) and input format (DIA-NN).
The constructor functions create a \code{TalusDataSet} instance
from a matrix, abundance files generated by metrics from DIA-NN,
and meta corresponding to the protein (peptide) annotation and
sample (run) annotation.
}
\author{
Chao-Jen Wong
}
