% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pca.R
\name{plot_pca}
\alias{plot_pca}
\title{Simple PCA plot for TalusDataSet and TalusDataSetList}
\usage{
plot_pca(object, ...)
}
\arguments{
\item{tds}{a \code{TalusDataSet} or \code{TalusDataSetList} object storing Talus data}

\item{top_n}{number of top proteins to use for principal components, selected by highest row variance}

\item{color_by}{a character string in colData to use for coloring}
}
\description{
This simple PC1 vs PC2 PCA helps to check for batch effects and the like.
}
