% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pca.R
\name{plot_pca}
\alias{plot_pca}
\alias{plot_pca,TalusDataSetList-method}
\alias{plot_pca,TalusDataSetList}
\alias{plot_pca,TalusDataSet-method}
\alias{plot_pca,TalusDataSet}
\title{Simple PCA scatter plot for TalusDataSet and TalusDataSetList}
\usage{
plot_pca(object, top_n = 500, color_by)

\S4method{plot_pca}{TalusDataSetList}(object, top_n = 500, color_by)

\S4method{plot_pca}{TalusDataSet}(object, top_n = 500, color_by)
}
\arguments{
\item{object}{a \code{TalusDataSet} or \code{TalusDataSetList} object storing Talus data.}

\item{top_n}{number of top proteins to use for principal components, selected by highest row variance (default 500).}

\item{color_by}{column name in \code{colData(object)} for point coloring (default none).}
}
\value{
a \code{ggplot} object for \code{TalusDataSet} or \code{TalusDataSetList}
}
\description{
Creates a PC1 vs PC2 scatter to inspect sample clustering, batch effects, etc.
}
