% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_per_protein.R
\name{plot_per_protein}
\alias{plot_per_protein}
\alias{plot_per_protein,TalusDataSetList-method}
\alias{plot_per_protein,TalusDataSetList}
\alias{plot_per_protein,TalusDataSet-method}
\alias{plot_per_protein,TalusDataSet}
\title{Simple point plot to show per-protein aboundance}
\usage{
plot_per_protein(object, protein_id, group_by)

\S4method{plot_per_protein}{TalusDataSetList}(object, protein_id, group_by)

\S4method{plot_per_protein}{TalusDataSet}(object, protein_id, group_by)
}
\arguments{
\item{object}{a \code{TalusDataSet} or \code{TalusDataSetList} instance containing the transformed protein abundance.}

\item{protein_id}{Protein ID}

\item{group_by}{group by a column (factor) in the meta file when plotting}
}
\description{
Given a protein ID, plot the MS intensity signal
}
\author{
Chao-Jen Wong
}
