% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_talus.R
\name{read_talus}
\alias{read_talus}
\title{Import Talus files}
\usage{
read_talus(
  file,
  meta_file,
  which_proteinid = "Protein.Ids",
  which_fraction = "Frx",
  which_sequence = NA,
  which_run = "Run",
  remove_few_measurements = TRUE,
  split_by_fraction = TRUE,
  log_transform = "log2",
  rowname_repair = TRUE,
  intensity_group = "protein",
  metric = "DIA-NN"
)
}
\arguments{
\item{file}{a path to a mass-spectrum intensity signal matric file (TSV OR CSV) or a connection.}

\item{which_proteinid}{which column in the \code{file} input represents the protein id}

\item{which_sequence}{which column in the \code{file} input represents the sequence of the protein}

\item{which_run}{which column in the meta file represents the runs (sample) ID.}

\item{remove_few_measurements}{if \code{TRUE}, remove measurements that are absent 85\% across all samples.}

\item{split_by_fraction}{if \code{TRUE}, split the protein abundance file according to value indicating in the values of the \code{which_fraction} column in the meta file.}

\item{log_transform}{if \code{TRUE}, perform log2 transformation to the protein abundance}

\item{intensity_group}{a character string specifying the mass-spectrum intensity signal level. Default to "protein".}

\item{metric}{a character specifying the metric data. Default to "DIA-NN".}

\item{meta_data}{a path to the metadata file (CSV) for the runs (samples).}
}
\value{
a \code{TalusDataSetList} object if \code{split_by_fraction} is TRUE, otherwise return a \code{TalusDataSet} object.
}
\description{
Import Talus files that were processed by DIANN and convert to
SummarizedExperiment instance. Can split the data into three instancs
containing fractions in nuc, cyto, and plasm.
}
