% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/talus_limma.R
\name{talus_limma}
\alias{talus_limma}
\alias{talus_limma,TalusDataSetList-method}
\alias{talus_limma,TalusDataSetList}
\alias{talus_limma,TalusDataSet-method}
\alias{talus_limma,TalusDataSet}
\title{Differential analysis via limma for TalusDataSet objects}
\usage{
talus_limma(object, design = ~0 + Tx)

\S4method{talus_limma}{TalusDataSetList}(object, design = ~0 + Tx)

\S4method{talus_limma}{TalusDataSet}(object, design = ~0 + Tx)
}
\arguments{
\item{object}{a \code{TalusDataSet} or \code{TalusDataSetList} instance containing protein log-transform assays}

\item{desgin}{a formula to create a model matrix. Default to ~0 + Tx, where \code{Tx} is a factor from colData of \code{se}}
}
\value{
a \code{TalusResult} if \code{object} is a \code{TalusDataSet}; or a \code{TalusResultList} if \code{TalusDataSetList}.
}
\description{
Fit linear models and contrasts using \code{limma::lmFit} and \code{limma::eBayes}
on a \code{TalusDataSet} or each element of a \code{TalusDataSetList}.
}
\author{
Chao-Jen Wong
}
