% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/talus_row_t_welch.R
\name{talus_row_t_welch}
\alias{talus_row_t_welch}
\alias{talus_row_t_welch,TalusDataSetList-method}
\alias{talus_row_t_welch,TalusDataSetList}
\alias{talus_row_t_welch,TalusDataSet-method}
\alias{talus_row_t_welch,TalusDataSet}
\title{Per-protein t-Welch testing}
\usage{
talus_row_t_welch(object, design = ~0 + Tx)

\S4method{talus_row_t_welch}{TalusDataSetList}(object, design = ~0 + Tx)

\S4method{talus_row_t_welch}{TalusDataSet}(object, design = ~0 + Tx)
}
\arguments{
\item{object}{a \code{TalusDataSet} or \code{TalusDataSetList} instance with log-transform assays}

\item{design}{a formula to create a model matrix. Default to ~0 + Tx, where \code{Tx} is a factor from colData of \code{se}}
}
\value{
a data.frame with t-statistics append to rowData of \code{object}
}
\description{
A wrapping function of matrxiTests::row_t_welch().
}
\author{
Chao-Jen Wong
}
