% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/talus_row_t_welch.R
\name{talus_row_t_welch}
\alias{talus_row_t_welch}
\title{Per-protien t-Weltch testing}
\usage{
talus_row_t_welch(se, design = ~0 + Tx)
}
\arguments{
\item{se}{A \code{SummarizedExperiment} instance with log-transform assays}

\item{design}{A formula to create a model matrix. Default to ~0 + Tx, where \code{Tx} is a factor from colData of \code{se}}
}
\value{
Orginal \code{SummarizedExperiment} objects with t-statistics append to rowData of \code{se}
}
\description{
A wrapping of
SummarizedExperiment instances with three layers: nuc, cyto, and plasm.
}
\author{
Chao-Jen Wong
}
