% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{TalusDataSet-class}
\alias{TalusDataSet-class}
\alias{TalusDataSet}
\title{TalusDataSet object and constructors}
\usage{
TalusDataSet(
  assay_data,
  col_data,
  row_data,
  intensity_group = "protein",
  metric = "DIA-NN"
)
}
\arguments{
\item{assay_data}{a matrix input: a matrix of non-negative integers of MS signal levels}

\item{col_data}{for matrix input: a \code{DataFrame} or \code{data.frame} with at least a single column. Rows of col_data correspond to columns of assay_ata}

\item{intensity_group}{a character string specifying the levels of abundance, either protein or peptide}

\item{matric}{a character string specifying the metric that generates the mass-spectrum signal values}
}
\description{
\code{TalusDataSet} is a subclass of \code{SummerizedExperiment} used
to store the input protein abundance values and addition annotation
of intensity group (protein or peptide) and input format (DIA-NN)
analysis of differential expression.
The constructor functions create a \code{TalusDataSet} instance
from a matrix, abundance files generated by metrics from DIA-NN,
and meta corresponding to the protein (peptide) annotation and
sample (run) annotation.
}
