% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{TalusDataSetList-class}
\alias{TalusDataSetList-class}
\alias{TalusDataSetList}
\title{TalusDataSetList-class and constructor}
\usage{
TalusDataSetList(lst)
}
\arguments{
\item{lst}{list of TalusDataSet objects}
}
\value{
a validated TalusDataSetList
}
\description{
A \code{TalusDataSetList} is an S4 \code{\linkS4class{SimpleList}} containing
one or more \code{\linkS4class{TalusDataSet}} objects.
}
\author{
Chao-Jen Wong
}
