% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{TalusDataSetList-class}
\alias{TalusDataSetList-class}
\alias{TalusDataSetList}
\title{TalusDataSetList object and constructors}
\usage{
TalusDataSetList(lst)
}
\arguments{
\item{lst}{A \code{list} of \code{TalusDataSet} objects.}
}
\value{
a \code{TalusDataSetList} object.
}
\description{
\code{TalusDataSetList} is a \code{SimpleList} object of \code{TalusDataSet}. The constructor convert a list of \code{TalusDataSet} objects to a \code{SimpleList} object.
}
\author{
Chao-Jen Wong
}
